Cleanup a few semle warnings (#1911)
Cleanup a few semle warnings.
Pass structure rather than 3 different parameters to locprint().

+++ db/osqlsession.c
@@ -451,9 +451,9 @@ void osql_sess_reqlogquery(osql_sess_t *sess, struct reqlogger *reqlog)
         reqlog, REQL_INFO,
         "rqid %s node %s sec %ld rtrs %d queuetime=%" PRId64 "ms \"%s\"\n",
         sess->rqid == OSQL_RQID_USE_UUID ? us : rqid,
-        (sess->offhost ? sess->offhost : ""),
-        (int)(sess->end - sess->initstart), reqlog_get_retries(reqlog),
-        reqlog_get_queue_time(reqlog) / 1000, sess->sql ? sess->sql : "()");
+        (sess->offhost ? sess->offhost : ""), (sess->end - sess->initstart),
+        reqlog_get_retries(reqlog), reqlog_get_queue_time(reqlog) / 1000,
+        sess->sql ? sess->sql : "()");
 }
 
 /**
